+++ moose/src/compilation/lowering.rs
